whatsapp_bot/requirements.txt
# FastAPI + PyWA WhatsApp Bot Dependencies

# Core framework and bot library
pywa[fastapi]==3.0.0
fastapi==0.115.5
uvicorn[standard]==0.32.0

# Environment and configuration
python-dotenv==1.0.1
pydantic==2.9.2

# HTTP client for external API calls
httpx==0.27.2
requests==2.32.3

# Optional: Database support (uncomment if needed)
# sqlalchemy==2.0.36
# databases[postgresql]==0.9.0

# Optional: Redis for caching (uncomment if needed)  
# redis==5.2.0

# Optional: AI/ML features (uncomment if needed)
# openai==1.54.4
# google-cloud-translate==3.18.0

# Development dependencies (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
